+++ Acciones/Controllers/SubcategoriasController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Acciones.Models;
+using System.Security.Claims;
 
 namespace Acciones.Controllers
 {
@@ -21,7 +22,12 @@ namespace Acciones.Controllers
         // GET: Subcategorias
         public async Task<IActionResult> Index()
         {
-            var dbAccionesContext = _context.Subcategorias.Include(s => s.Categoria);
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
+            }
+            //var dbAccionesContext = _context.Subcategorias.Include(s => s.Categoria).OrderBy(s => s.CategoriaId);
+            var dbAccionesContext = _context.Subcategorias.Include(c => c.Categoria).OrderBy(c=>c.Categoria.Descripcion).ThenBy(sc=>sc.Descripcion);
             return View(await dbAccionesContext.ToListAsync());
         }
 
@@ -31,6 +37,11 @@ namespace Acciones.Controllers
             if (id == null)
             {
                 return NotFound();
+            }
+
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
             }
 
             var subcategoria = await _context.Subcategorias
@@ -47,6 +58,11 @@ namespace Acciones.Controllers
         // GET: Subcategorias/Create
         public IActionResult Create()
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
+            }
+
             ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "CategoriaId");
             return View();
         }
@@ -64,6 +80,12 @@ namespace Acciones.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
+            }
+
             ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "CategoriaId", subcategoria.CategoriaId);
             return View(subcategoria);
         }
@@ -74,6 +96,11 @@ namespace Acciones.Controllers
             if (id == null)
             {
                 return NotFound();
+            }
+
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
             }
 
             var subcategoria = await _context.Subcategorias.FindAsync(id);
@@ -97,6 +124,16 @@ namespace Acciones.Controllers
                 return NotFound();
             }
 
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
+            }
+
+            //subcategoria.FechaUltimoCambio = DateTime.Now;
+            //string idU = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            //var idUsr = Int32.Parse(idU);
+            //subcategoria.UsuarioRegistroId = idUsr;
+
             if (ModelState.IsValid)
             {
                 try
@@ -118,7 +155,8 @@ namespace Acciones.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["CategoriaId"] = new SelectList(_context.Categorias, "CategoriaId", "CategoriaId", subcategoria.CategoriaId);
-            return View(subcategoria);
+            //return View(subcategoria);
+            return RedirectToAction("Index", "Categorias");
         }
 
         // GET: Subcategorias/Delete/5
@@ -127,6 +165,11 @@ namespace Acciones.Controllers
             if (id == null)
             {
                 return NotFound();
+            }
+
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
             }
 
             var subcategoria = await _context.Subcategorias
@@ -145,6 +188,11 @@ namespace Acciones.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Redirect("/Apoyo/Home/Inicio");
+            }
+
             var subcategoria = await _context.Subcategorias.FindAsync(id);
             _context.Subcategorias.Remove(subcategoria);
             await _context.SaveChangesAsync();
@@ -162,9 +210,90 @@ namespace Acciones.Controllers
         public JsonResult BusquedaSubcategoria(string term, int categoriaId)
         {
             {
-                var nombresSubcategorias = _context.Subcategorias.Where(sc => sc.Descripcion .Contains(term) && sc.CategoriaId == categoriaId).OrderBy(sc=>sc.Descripcion).Select(c => new { aNombreSubcategoria = c.Descripcion, id = c.SubcategoriasId }).ToList();
+                var nombresSubcategorias = _context.Subcategorias.Where(sc => sc.Descripcion.Contains(term) && sc.CategoriaId == categoriaId).OrderBy(sc => sc.Descripcion).Select(c => new { aNombreSubcategoria = c.Descripcion, id = c.SubcategoriasId }).ToList();
                 return Json(nombresSubcategorias);
             }
         }
+
+        /****************************************************************************
+                Obtiene Listado de Subcategorias correspondientes a una Categoria ID
+         *****************************************************************************/
+        [HttpGet]
+        public List<Subcategoria> ObtieneListaSubCat(int categoriaId)
+        {
+            List<Subcategoria> ListaObjetos = new List<Subcategoria>();
+
+            ListaObjetos = _context.Subcategorias.Where(sc => sc.CategoriaId == categoriaId).OrderBy(sc=>sc.Descripcion).ToList();
+            return ListaObjetos;
+        }
+
+
+        /****************************************************************************
+                     Actualiza SubCategorias
+         *****************************************************************************/
+        [HttpPut]
+        public int ActualizaSubCategoria(int subCategoriaId, string descSubCategoria, int estatus)
+        {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return 0;
+            }
+
+            Subcategoria subCategoria = _context.Subcategorias.Where(sc => sc.SubcategoriasId == subCategoriaId).FirstOrDefault();
+
+            subCategoria.Descripcion = descSubCategoria;
+            subCategoria.Estatus = estatus;
+            subCategoria.FechaUltimoCambio = DateTime.Now;
+            string idU = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var idUsr = Int32.Parse(idU);
+            subCategoria.UsuarioRegistroId = idUsr;
+
+            try
+            {
+                _context.Update(subCategoria);
+                _context.SaveChanges();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!SubcategoriaExists(subCategoria.SubcategoriasId))
+                {
+                    return 0;
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return 1;
+        }
+
+        /****************************************************************************
+                Elimina Categorias
+         *****************************************************************************/
+        [HttpDelete]
+        public int EliminaSubCategoria(int subCategoriaId)
+        {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return 0;
+            }
+
+            try
+            {
+                var subcategoria = _context.Subcategorias.Find(subCategoriaId);
+                _context.Subcategorias.Remove(subcategoria);
+                _context.SaveChangesAsync();
+
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                return 0;
+            }
+
+            return 1;
+        }
+
+
+
     }
 }
